Added Pause Shape
